move service daemon codes out of launch.py
